+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -71,6 +71,7 @@ class GeckoEngine(
             set(value) {
                 value?.let {
                     runtime.settings.trackingProtectionCategories = it.categories
+                    defaultSettings?.trackingProtectionPolicy = value
                 }
             }